requirements.txt
exceptiongroup==1.1.3
iniconfig==2.0.0
mysql-connector-python==8.2.0
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
pytest==7.4.3
tomli==2.0.1
